# bedelia-rest/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Bedelia UTEC",
        "contact": {
            "name": "Support Team",
            "email": "support@example.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/areasEstudio/{id}": {
            "get": {
                "tags": [
                    "Areas de estudio"
                ],
                "operationId": "App\\Http\\Controllers\\AreaEstudioController::obtenerUno",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AreaEstudioDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/areasEstudio": {
            "get": {
                "tags": [
                    "Areas de estudio"
                ],
                "operationId": "App\\Http\\Controllers\\AreaEstudioController::obtenerLista",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AreaEstudioDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Areas de estudio"
                ],
                "operationId": "App\\Http\\Controllers\\AreaEstudioController::agregar",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AreaEstudioDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AreaEstudioDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/areasEstudio/{id}/cursos": {
            "get": {
                "tags": [
                    "Areas de estudio"
                ],
                "description": "Devuelve los cursos que pertenecen al area de estudio",
                "operationId": "App\\Http\\Controllers\\AreaEstudioController::obtenerCursosPertenecientesAUnArea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AreaEstudioDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/carreras/{id}": {
            "get": {
                "tags": [
                    "Carreras"
                ],
                "operationId": "App\\Http\\Controllers\\CarrerasController::obtenerUno",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CarreraDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/carreras/": {
            "post": {
                "tags": [
                    "Carreras"
                ],
                "operationId": "App\\Http\\Controllers\\CarrerasController::agregar",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarreraCreateDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CarreraDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/cursos/{id}": {
            "get": {
                "tags": [
                    "Cursos"
                ],
                "operationId": "App\\Http\\Controllers\\CursoController::obtenerUno",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CursoDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/cursos": {
            "get": {
                "tags": [
                    "Cursos"
                ],
                "operationId": "App\\Http\\Controllers\\CursoController::obtenerLista",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CursoDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Cursos"
                ],
                "operationId": "App\\Http\\Controllers\\CursoController::agregar",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CursoDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CursoDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/edicionesCurso/{id}/clasesDictada": {
            "post": {
                "tags": [
                    "Ediciones Curso"
                ],
                "description": "Registra una nueva ClaseDictada y la asistencias de los estudiantes",
                "operationId": "App\\Http\\Controllers\\EdicionesCursoController::Agregar",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del edicion curso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClaseDictadaDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClaseDictadaDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/periodos": {
            "get": {
                "tags": [
                    "Periodos"
                ],
                "operationId": "App\\Http\\Controllers\\PeriodoLectivoController::obtenerLista",
                "responses": {
                    "200": {
                        "description": "Devuelve el ultimo anio lectivo ingresado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnioLectivoDTO"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Periodos"
                ],
                "operationId": "App\\Http\\Controllers\\PeriodoLectivoController::agregar",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AnioLectivoDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "No devuelve nada, ya es bastante que funcione xD"
                    }
                }
            }
        },
        "/postulantes/{id}": {
            "get": {
                "tags": [
                    "Postulante"
                ],
                "description": "devuelve una postulacion especifica, incluyendo la persona asociada y su direccion",
                "operationId": "App\\Http\\Controllers\\PostulantesController::obtenerUno",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la postulacion",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostulanteDTO"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Postulante"
                ],
                "description": "Elimina la postulacion especifica",
                "operationId": "App\\Http\\Controllers\\PostulantesController::rechazar",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la postulacion",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/postulantes": {
            "post": {
                "tags": [
                    "Postulante"
                ],
                "description": "Registra en el sistema los datos de un postulante",
                "operationId": "App\\Http\\Controllers\\PostulantesController::agregar",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostulanteDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostulanteDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/sedes/{id}": {
            "get": {
                "tags": [
                    "Sedes"
                ],
                "operationId": "App\\Http\\Controllers\\SedesController::obtenerUno",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la sede",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve datos de la sede",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SedeDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/sedes": {
            "post": {
                "tags": [
                    "Sedes"
                ],
                "operationId": "App\\Http\\Controllers\\SedesController::agregar",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SedeDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Devuelve datos de la sede creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SedeDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/tiposCurso/{id}": {
            "get": {
                "tags": [
                    "Tipos de curso"
                ],
                "operationId": "App\\Http\\Controllers\\TipoCursoController::obtenerUno",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TipoCursoDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tiposCurso": {
            "get": {
                "tags": [
                    "Tipos de curso"
                ],
                "operationId": "App\\Http\\Controllers\\TipoCursoController::obtenerLista",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TipoCursoDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tipos de curso"
                ],
                "operationId": "App\\Http\\Controllers\\TipoCursoController::agregar",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TipoCursoDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TipoCursoDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/usuarios/{ci}": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "operationId": "App\\Http\\Controllers\\UsuarioController::obtenerUno",
                "parameters": [
                    {
                        "name": "ci",
                        "in": "path",
                        "description": "CI del usuario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve datos del usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsuarioDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/usuarios": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "operationId": "App\\Http\\Controllers\\UsuarioController::agregar",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UsuarioDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Devuelve datos del usuario creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsuarioDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UsuarioDTO": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "contrasenia": {
                        "type": "string"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "persona": {
                        "$ref": "#/components/schemas/PersonaDTO"
                    }
                },
                "type": "object"
            },
            "SedeDTO": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "telefono": {
                        "type": "string"
                    },
                    "direccion": {
                        "$ref": "#/components/schemas/DireccionDTO"
                    }
                },
                "type": "object"
            },
            "TipoCursoDTO": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "tipo": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "AreaEstudioDTO": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "area": {
                        "type": "string"
                    },
                    "creditos": {
                        "description": "Depende de la carrera",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "CursoDTO": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "descripcion": {
                        "type": "string"
                    },
                    "max_inasistencias": {
                        "type": "integer"
                    },
                    "cant_creditos": {
                        "type": "integer"
                    },
                    "cant_clases": {
                        "type": "integer"
                    },
                    "semestre": {
                        "description": "Depende de la carrera",
                        "type": "integer"
                    },
                    "optativo": {
                        "description": "Depende de la carrera",
                        "type": "integer"
                    },
                    "area_estudio": {
                        "$ref": "#/components/schemas/AreaEstudioDTO"
                    },
                    "tipo_curso": {
                        "$ref": "#/components/schemas/TipoCursoDTO"
                    }
                },
                "type": "object"
            },
...
            "CarreraDTO": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "descripcion": {
                        "type": "string"
                    },
                    "cant_semestres": {
                        "type": "integer"
                    },
                    "areas_estudio": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AreaEstudioDTO"
                        }
                    },
                    "sedes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SedeDTO"
                        }
                    }
                },
                "type": "object"
            },
            "PreviaDTO": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "carrera_id": {
                        "type": "integer"
                    },
                    "curso_id": {
                        "type": "integer"
                    },
                    "curso_id_previa": {
                        "type": "integer"
                    },
                    "tipo": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "CarreraCreateDTO": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "descripcion": {
                        "type": "string"
                    },
                    "cant_semestres": {
                        "type": "integer"
                    },
                    "sedes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SedeDTO"
                        }
                    },
                    "areas_estudio": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AreaEstudioDTO"
                        }
                    },
                    "previas": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PreviaDTO"
                        }
                    },
                    "cursos": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "semestre": {
                                    "type": "integer"
                                },
                                "optativo": {
                                    "type": "boolean"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "PostulanteDTO": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "img_ci": {
                        "type": "string"
                    },
                    "img_escolaridad": {
                        "type": "string"
                    },
                    "img_carne_salud": {
                        "type": "string"
                    },
                    "sede": {
                        "$ref": "#/components/schemas/SedeDTO"
                    },
                    "carrera": {
                        "$ref": "#/components/schemas/CarreraDTO"
                    },
                    "persona": {
                        "$ref": "#/components/schemas/PersonaDTO"
                    }
                },
                "type": "object"
            },
...
            "ClaseDictadaDTO": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "fecha": {
                        "type": "string",
                        "readOnly": true
                    },
                    "curso_id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "edicion_curso_id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "lista": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "ciEstudiante": {
                                    "type": "string"
                                },
                                "nombre": {
                                    "type": "string",
                                    "readOnly": true
                                },
                                "apellido": {
                                    "type": "string",
                                    "readOnly": true
                                },
                                "asistencia": {
                                    "type": "number",
                                    "writeOnly": true
                                },
                                "total_asistencias": {
                                    "type": "number",
                                    "readOnly": true
                                },
                                "total_llegadas_tarde": {
                                    "type": "number",
                                    "readOnly": true
                                },
                                "total_inasistencias": {
                                    "type": "number",
                                    "readOnly": true
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "description": "Token de autenticacion conformato como '<token>'",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}